access control implemented

+++ app/Http/Controllers/PostController.php
@@ -8,6 +8,15 @@ use App\Models\Post;
 class PostController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -75,6 +84,11 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !== $post->user->id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -110,6 +124,12 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !== $post->user->id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
